+++ src/main/java/com/sotatek/cardano/common/entity/ParamProposal.java
@@ -5,7 +5,7 @@ import com.sotatek.cardano.common.validation.Hash32Type;
 import com.sotatek.cardano.common.validation.Lovelace;
 import com.sotatek.cardano.common.validation.Word31Type;
 import com.sotatek.cardano.common.validation.Word64Type;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Objects;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -49,47 +49,47 @@ public class ParamProposal extends BaseEntity {
   @Column(name = "min_fee_a", precision = 20)
   @Word64Type
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal minFeeA;
+  private BigInteger minFeeA;
 
   @Column(name = "min_fee_b", precision = 20)
   @Word64Type
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal minFeeB;
+  private BigInteger minFeeB;
 
   @Column(name = "max_block_size", precision = 20)
   @Word64Type
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal maxBlockSize;
+  private BigInteger maxBlockSize;
 
   @Column(name = "max_tx_size", precision = 20)
   @Word64Type
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal maxTxSize;
+  private BigInteger maxTxSize;
 
   @Column(name = "max_bh_size", precision = 20)
   @Word64Type
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal maxBhSize;
+  private BigInteger maxBhSize;
 
   @Column(name = "key_deposit")
   @Lovelace
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal keyDeposit;
+  private BigInteger keyDeposit;
 
   @Column(name = "pool_deposit", precision = 20)
   @Lovelace
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal poolDeposit;
+  private BigInteger poolDeposit;
 
   @Column(name = "max_epoch", precision = 20)
   @Word64Type
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal maxEpoch;
+  private BigInteger maxEpoch;
 
   @Column(name = "optimal_pool_count", precision = 20)
   @Word64Type
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal optimalPoolCount;
+  private BigInteger optimalPoolCount;
 
   @Column(name = "influence")
   private Double influence;
@@ -118,12 +118,12 @@ public class ParamProposal extends BaseEntity {
   @Column(name = "min_utxo_value", precision = 20)
   @Lovelace
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal minUtxoValue;
+  private BigInteger minUtxoValue;
 
   @Column(name = "min_pool_cost", precision = 20)
   @Lovelace
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal minPoolCost;
+  private BigInteger minPoolCost;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @OnDelete(action = OnDeleteAction.CASCADE)
@@ -141,27 +141,27 @@ public class ParamProposal extends BaseEntity {
   @Column(name = "max_tx_ex_mem", precision = 20)
   @Word64Type
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal maxTxExMem;
+  private BigInteger maxTxExMem;
 
   @Column(name = "max_tx_ex_steps", precision = 20)
   @Word64Type
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal maxTxExSteps;
+  private BigInteger maxTxExSteps;
 
   @Column(name = "max_block_ex_mem", precision = 20)
   @Word64Type
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal maxBlockExMem;
+  private BigInteger maxBlockExMem;
 
   @Column(name = "max_block_ex_steps", precision = 20)
   @Word64Type
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal maxBlockExSteps;
+  private BigInteger maxBlockExSteps;
 
   @Column(name = "max_val_size", precision = 20)
   @Word64Type
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal maxValSize;
+  private BigInteger maxValSize;
 
   @Column(name = "collateral_percent")
   @Word31Type
@@ -181,7 +181,7 @@ public class ParamProposal extends BaseEntity {
   @Column(name = "coins_per_utxo_size")
   @Lovelace
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal coinsPerUtxoSize;
+  private BigInteger coinsPerUtxoSize;
 
   @Override
   public boolean equals(Object o) {
