+++ src/main/java/com/sotatek/cardano/common/entity/PoolUpdate.java
@@ -2,7 +2,7 @@ package com.sotatek.cardano.common.entity;
 
 import com.sotatek.cardano.common.validation.Hash32Type;
 import com.sotatek.cardano.common.validation.Lovelace;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Objects;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -52,7 +52,7 @@ public class PoolUpdate extends BaseEntity {
   @Column(name = "pledge", nullable = false, precision = 20)
   @Lovelace
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal pledge;
+  private BigInteger pledge;
 
   @ManyToOne(fetch = FetchType.LAZY, optional = false)
   @OnDelete(action = OnDeleteAction.CASCADE)
@@ -77,7 +77,7 @@ public class PoolUpdate extends BaseEntity {
   @Column(name = "fixed_cost", nullable = false, precision = 20)
   @Lovelace
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal fixedCost;
+  private BigInteger fixedCost;
 
   @ManyToOne(fetch = FetchType.LAZY, optional = false)
   @OnDelete(action = OnDeleteAction.CASCADE)
